Canvas refactoring.
